requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.10
djangorestframework==3.14.0
idna==3.6
pillow==10.2.0
pytz==2024.1
requests==2.31.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.0
scikit-learn==1.4.1.post1
opencv-python==4.8.0.74
joblib==1.3.2